Kimi-K3: keep the KDA tensors the checkpoint stores in fp32 in fp32
Kimi-K3 checkpoints five KDA/router tensors in fp32 where Kimi-Linear-48B
checkpoints them in bf16: q/k/v_conv1d.weight, o_norm.weight, and
block_sparse_moe.gate.e_score_correction_bias (A_log and dt_bias are fp32 in
both). Those are the recurrent-state and expert-selection tensors, i.e. the
ones where rounding to bf16 is least obviously harmless -- the correction bias
decides the top-16 cut.

Swept all 497,052 released text tensors against the parameters they load
into. The router bias was already fine: DeepSeekV3Router builds it at
router_bias_dtype, which defaults to fp32 and the Kimi MoE block does not
override, so the checkpoint value reaches get_topk_indices intact.

The KDA sublayer had two real narrowings, both in _params rather than in the
load itself (jax_array_from_reshaped_torch converts with t2j and keeps the
checkpoint dtype):

  * o_norm.weight was cast to the model dtype and then widened straight back
    to fp32 inside the gated RMSNorm, which computes in fp32. Pure loss; the
    cast is gone and the parameter is declared fp32.
  * the short-conv weights are still met at the activation dtype, because the
    ragged conv's prefill branch is a lax.conv_general_dilated and that
    rejects mixed dtypes. The parameter is declared fp32 anyway so a dummy
    init matches the checkpoint and vLLM's KDA layer
    (params_dtype=torch.float32 on its conv1d), and the one place the extra
    precision is dropped is now named in a comment instead of being implicit.
    The other two conv branches -- the decode einsum and the loop that
    already accumulates in fp32 -- would take an fp32 weight directly, so
    running the whole conv in fp32 is a compute-dtype question for
    ragged_conv1d, not a loading one.

Numerics are unchanged for both released checkpoints: widening a bf16
checkpoint value to fp32 is exact, and the conv weight still reaches the
kernel at the model dtype.

+++ tpu_inference/models/jax/kimi_k3.py
@@ -465,8 +465,14 @@ class KimiDeltaAttention(JaxModule):
             )
 
         def _conv() -> _RawWeight:
+            # fp32, not the model dtype: Kimi-K3 checkpoints the short-conv
+            # weights in fp32 (Kimi-Linear-48B checkpoints them in bf16), and
+            # a bf16 parameter would round them away at load. vLLM's KDA layer
+            # holds them in fp32 for the same reason
+            # (`params_dtype=torch.float32` on its conv1d). See `_params` for
+            # where the compute dtype is applied.
             return _RawWeight(shape=(HP, 1, self.conv_kernel_size),
-                              dtype=self.dtype,
+                              dtype=jnp.float32,
                               rngs=rngs,
                               sharding=P(),
                               random_init=self.random_init)
@@ -496,12 +502,18 @@ class KimiDeltaAttention(JaxModule):
                                     dtype=jnp.float32,
                                     sharding=P(),
                                     random_init=self.random_init)
+        # fp32 for the same reason as the conv weights: Kimi-K3 checkpoints
+        # `o_norm.weight` in fp32, and the gated RMSNorm that consumes it runs
+        # in fp32 anyway, so a bf16 parameter would round the checkpoint value
+        # away for nothing. (Kimi-Linear-48B checkpoints it in bf16; widening
+        # that is exact.) Only the weight is used -- the norm math itself
+        # lives in `kda_attention._gated_rmsnorm`.
         self.o_norm = JaxRmsNorm(K,
                                  epsilon=self.rms_norm_eps,
                                  scale_init=nnx.with_partitioning(
                                      init_fn, (None, )),
-                                 param_dtype=self.dtype,
-                                 dtype=self.dtype,
+                                 param_dtype=jnp.float32,
+                                 dtype=jnp.float32,
                                  quant_config=None,
                                  prefix=self.prefix + ".o_norm",
                                  rngs=rngs)
@@ -516,8 +528,7 @@ class KimiDeltaAttention(JaxModule):
         def _t(layer: JaxEinsum) -> jax.Array:
             # Weights keep their checkpoint dtype; the layer decides the
             # compute dtype (they differ when a bf16 checkpoint is run in fp32
-            # for parity checks, and `lax.conv_general_dilated` rejects mixed
-            # dtypes).
+            # for parity checks).
             return layer.weight.value.astype(self.dtype).T
 
         gate_kwargs = {}
@@ -530,6 +541,14 @@ class KimiDeltaAttention(JaxModule):
             q_proj=_t(self.q_proj),
             k_proj=_t(self.k_proj),
             v_proj=_t(self.v_proj),
+            # The conv weights are held in fp32 (see `_conv`) but the ragged
+            # conv's prefill branch is a `lax.conv_general_dilated`, which
+            # rejects mixed dtypes, so the weight meets the activations at the
+            # model dtype. This is the one place the checkpoint's extra conv
+            # precision is dropped; the other two branches (the decode einsum
+            # and the accumulate-in-fp32 loop) would take fp32 directly, so
+            # running the whole conv in fp32 is a compute-dtype decision for
+            # `ragged_conv1d`, not a loading one.
             q_conv_weight=self.q_conv1d.weight.value.astype(self.dtype),
             k_conv_weight=self.k_conv1d.weight.value.astype(self.dtype),
             v_conv_weight=self.v_conv1d.weight.value.astype(self.dtype),
@@ -540,7 +559,9 @@ class KimiDeltaAttention(JaxModule):
             f_a_proj=_t(self.f_a_proj),
             f_b_proj=_t(self.f_b_proj),
             b_proj=_t(self.b_proj),
-            o_norm_weight=self.o_norm.weight.value.astype(self.dtype),
+            # Not cast: the gated RMSNorm runs in fp32 and the parameter is
+            # already fp32, so the checkpoint value reaches the math intact.
+            o_norm_weight=self.o_norm.weight.value,
             o_proj=_t(self.o_proj),
             **gate_kwargs)
